functions/get_ticket_resolutions/requirements.txt
functions-framework==3.*
google-cloud-bigquery
pandas
pyarrow
db-dtypes
